Add files via upload
Here is the current SVM ML model, it is trained on a 60/20/20 split to test which hyperparemeters work best. Out of the selkection, it will take the one with the highest accuracy. The highest accuracy will then be trained and tested to see the breakdown in accuracy and percision. 
The next splits will be added to be compared with the other ML models as well.